refactor: forgot password file structure
