+++ app/javascript/authentication/ForgotPassword.module.scss
