+++ app/javascript/authentication/ForgotPasswordForm.tsx
